refactor(mobile): improve signature service architecture and cleanup coverage configuration
## Architecture Improvements

- **Move TypeScript interfaces to dedicated types directory:**
  - Create `src/services/signature/types/` directory structure
  - Move `SignatureStrategy` and `SignatureConfig` interfaces from strategies/ to types/interfaces.ts
  - Add proper barrel export with coverage exclusion

- **Clean separation of concerns:**
  - `/types/` directory: Pure TypeScript interface definitions
  - `/strategies/` directory: Concrete implementations only
  - Better follows TypeScript architectural best practices

- **Update all import references (7 files):**
  - RegularWalletStrategy.ts: Import from ../types
  - SafeWalletStrategy.ts: Import from ../types
  - SignatureStrategyFactory.ts: Import from ../types
  - SignatureService.test.ts: Import from ./strategies (re-exported)
  - strategies/index.ts: Re-export types from ../types

## Coverage Configuration Improvements

- **Remove interface-only file from coverage scope:**
  - Delete SignatureStrategy.ts (interface definitions only)
  - Delete SignatureStrategy.test.ts (testing TypeScript compiler compliance)
  - Add `!src/**/types/**` exclusion pattern for interface directories
  - Keep `!src/**/index.ts` exclusion for barrel exports

- **Remove redundant istanbul ignore comments:**
  - Clean up 15+ barrel export files (index.ts)
  - Comments were redundant since Jest config excludes index.ts files
  - Improved code consistency and cleanliness

## Benefits Achieved

- ✅ **No more 0% coverage confusion** from interface-only files
- ✅ **Cleaner architecture** following TypeScript best practices
- ✅ **Better separation** between types and implementations
- ✅ **All tests passing** with updated import structure
- ✅ **Reduced code clutter** with unnecessary coverage comments removed
- ✅ **Improved maintainability** for future interface additions

Quality assurance: TypeScript ✅ Formatting ✅ Linting ✅
