+++ apps/mobile/src/components/index.ts
@@ -1,4 +1,3 @@
-/* istanbul ignore file */
 // Component exports for mobile app
 export { AuthProgressIndicator } from './AuthProgressIndicator'
 export { LoadingSpinner } from './LoadingSpinner'
