+++ apps/mobile/src/hooks/auth/index.ts
@@ -1,4 +1,3 @@
-/* istanbul ignore file */
 // Authentication domain hooks
 export { useAuthentication } from './useAuthentication'
 export type { Authentication } from './useAuthentication'
