+++ apps/mobile/src/hooks/index.ts
@@ -1,4 +1,3 @@
-/* istanbul ignore file */
 // Hook domain re-exports for cleaner imports
 // Example: import { useAuthentication, useWalletToasts } from '../hooks'
