+++ apps/mobile/src/hooks/ui/index.ts
@@ -1,3 +1,2 @@
-/* istanbul ignore file */
 // UI/UX domain hooks
 export { useGlobalErrorHandler } from './useGlobalErrorHandler'
