+++ apps/mobile/src/hooks/wallet/index.ts
@@ -1,4 +1,3 @@
-/* istanbul ignore file */
 // Wallet domain hooks
 export { useWalletConnectionTrigger } from './useWalletConnectionTrigger'
