+++ apps/mobile/src/services/authentication/index.ts
@@ -1,4 +1,3 @@
-/* istanbul ignore file */
 // Authentication service exports
 export { AuthenticationLock, AuthenticationOrchestrator } from './AuthenticationOrchestrator'
