+++ apps/mobile/src/services/index.ts
@@ -1,4 +1,3 @@
-/* istanbul ignore file */
 // Main services exports for backward compatibility and easy imports
 
 // Authentication services
