+++ apps/mobile/src/services/signature/strategies/RegularWalletStrategy.ts
@@ -1,6 +1,6 @@
 import type { Connector } from 'wagmi'
 import type { SignatureFunctions, SignatureRequest, SignatureResult } from '@superpool/types'
-import type { SignatureStrategy } from './SignatureStrategy'
+import type { SignatureStrategy } from '../types'
 import { SignatureUtils } from './SignatureUtils'
 
 /**
