+++ apps/mobile/src/services/signature/strategies/index.ts
@@ -1,7 +1,7 @@
-/* istanbul ignore file */
 // Signature strategy modules for clean separation of concerns and wallet-specific logic
 export { RegularWalletStrategy } from './RegularWalletStrategy'
 export { SafeWalletStrategy } from './SafeWalletStrategy'
-export { SignatureConfig, SignatureStrategy } from './SignatureStrategy'
 export { SignatureStrategyFactory } from './SignatureStrategyFactory'
 export { SignatureUtils } from './SignatureUtils'
+// Re-export types from types directory
+export type { SignatureConfig, SignatureStrategy } from '../types'
