+++ apps/mobile/src/services/signature/types/index.ts
@@ -1,0 +1,5 @@
+/**
+ * TypeScript interface definitions for signature services
+ * These interfaces define contracts but don't generate runtime code
+ */
+export type { SignatureConfig, SignatureStrategy } from './interfaces'
