+++ apps/mobile/src/services/signature/types/interfaces.ts
