+++ apps/mobile/src/services/utils/index.ts
@@ -1,4 +1,3 @@
-/* istanbul ignore file */
 // Service utilities for common operations across authentication and signature services
 // Provides centralized, reusable utility functions with consistent patterns
